+++ app/controllers/api/v1/qcs_controller.rb
@@ -42,7 +42,7 @@ module Api
         @new_qc = @qc.deep_clone include: %i[qc_attacks qc_charms qc_merits],
                                  except: %i[chronicle_id sort_order chronicle_sort_order pinned hidden public]
 
-        @new_qc.name = @new_qc.name + " (Duplicate)"
+        @new_qc.name = @new_qc.name + ' (Duplicate)'
         @new_qc.player = current_player
 
         if @new_qc.save