+++ app/models/concerns/health_levels.rb
@@ -14,5 +14,13 @@ module HealthLevels
               :damage_lethal,
               :damage_aggravated,
               numericality: { greater_than_or_equal_to: 0 }
+
+    def total_health_levels
+      health_level_0s +
+        health_level_1s +
+        health_level_2s +
+        health_level_4s +
+        health_level_incap
+    end
   end
 end